working slow ahh training

+++ src/main/java/io/github/betterclient/ai/neural/Neuron.java
@@ -4,18 +4,18 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Neuron {
-    public Map<Neuron, Double> connectionWeights = new HashMap<>();
-    public Map<Neuron, Double> costGradientW = new HashMap<>();
-    public double bias;
-    public double costGradientB;
+    public Map<Neuron, Float> connectionWeights = new HashMap<>();
+    public Map<Neuron, Float> costGradientW = new HashMap<>();
+    public float bias;
+    public float costGradientB;
 
-    public Neuron(double bias) {
+    public Neuron(float bias) {
         this.bias = bias;
     }
 
     public void initConnections(NeuralLayer layer) {
         for (Neuron neuron : layer.neurons) {
-            connectionWeights.put(neuron, 0D);
+            connectionWeights.put(neuron, 0F);
         }
     }
 }
